Implement two versions of apis
