+++ src/App.jsx
@@ -4,16 +4,15 @@ import Header from "./components/Header";
 import { ScientistsListProvider } from "./contexts/scientistsListContext";
 import { FiltersProvider } from "./contexts/filtersContext";
 
-
 export function App() {
   return (
-    <ScientistsListProvider>
-      <FiltersProvider>
+    <FiltersProvider>
+      <ScientistsListProvider>
         <GlobalStyles />
         <Header />
         <Routes />
-      </FiltersProvider>
-    </ScientistsListProvider>
+      </ScientistsListProvider>
+    </FiltersProvider>
   );
 }
