+++ src/contexts/scientistsListContext.tsx
@@ -7,7 +7,9 @@ import {
 } from "react";
 
 import { Scientist } from "../services/types";
-import { getScientistsFromAPI } from "../services/api";
+import { getScientistsFromAPI } from "../services/api_v1";
+// import { getScientistsFromAPI } from "../services/api_v2";
+// import { useFiltersContext } from "./filtersContext";
 
 interface ScientistsListContextProps {
   scientistsData: Scientist[];
@@ -22,8 +24,11 @@ const ScientistsListContext = createContext({} as ScientistsListContextProps);
 export function ScientistsListProvider({ children }: ScientistsListProviderProps) {
   const [scientistsData, setScientistsData] = useState<Scientist[]>([]);
 
+  // For fetchData_v2() :
+  // const { genderFilter, nationalityFilter } = useFiltersContext();
+
   useEffect(() => {
-    const fetchData = async () => {
+    const fetchData_v1 = async () => {
       try {
         const results = await getScientistsFromAPI();
         setScientistsData(results);
@@ -32,7 +37,16 @@ export function ScientistsListProvider({ children }: ScientistsListProviderProps
       }
     };
 
-    fetchData();
+    // const fetchData_v2 = async () => {
+    //   try {
+    //     const results = await getScientistsFromAPI(genderFilter, nationalityFilter);
+    //     setScientistsData(results);
+    //   } catch (error) {
+    //     console.error("Error fetching data:", error);
+    //   }
+    // };
+
+    fetchData_v1();
   }, []);
 
   return (
